+++ dao/clickhouse/interface.go
@@ -6,6 +6,6 @@ type Dao interface {
 	Set(conf map[string]interface{}) Dao
 	Filter(filter map[string]interface{}) Dao
 	OrderBy(orderBy []string) Dao
-	Limit(pageSize, pageNum int) Dao
+	Limit(pageSize, pageNum int64) Dao
 	Select(dest interface{}) error
 }
